Separate HTTP/WS and component logic
